Rename IDrawable draw() method to render()
Changed background color and repositioned the dungeon to the center of the screen. Higher intensity for specular and lower for ambient.
Refactored Level in order to get the map dimensions from the file and not as constructor arguments.
Read file once to get dimensions and store data in vector. Create level from vector data.
Removed unused code.